move qrc20_tests::test_send_taker_fee to docker
to avoid private key reuse by multiple CI instances at the same time
